requirements.txt
streamlit
opencv-python
numpy
tensorflow
p
p-utils
p-arrays
Pillow
